writing-practice/requirements.txt
streamlit==1.32.0
openai>=1.12.0
pytesseract==0.3.10
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1 